Muevo Stage a su propio módulo
