+++ django_datajsonar/admin/__init__.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
-from django_datajsonar.models import Jurisdiction, Stage
+from django_datajsonar.models import Jurisdiction
+from django_datajsonar.models import Stage
 from . import synchronizer, data_json, metadata, node, repeatable_job, tasks
 
 admin.site.register(Jurisdiction)
