+++ django_datajsonar/models/__init__.py
@@ -3,6 +3,7 @@ from .utils import filepath, get_distribution_storage
 from .data_json import Catalog, Dataset, Distribution, Field
 from .metadata import Metadata
 from .tasks import ReadDataJsonTask, AbstractTask
-from .synchronizer import Synchronizer, Stage
+from .synchronizer import Synchronizer
+from .stage import Stage
 from .node import Node, NodeMetadata, NodeRegisterFile,\
     DatasetIndexingFile, Jurisdiction
